Unit tests restructure to reduce duplicated code.
