+++ src/test/kotlin/eu/europa/ec/eudi/openid4vp/internal/request/RequestAuthenticatorTest.kt
@@ -19,11 +19,7 @@ import com.nimbusds.jose.JOSEException
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.JWSHeader
 import com.nimbusds.jose.JWSVerifier
-import com.nimbusds.jose.jwk.Curve
-import com.nimbusds.jose.jwk.ECKey
 import com.nimbusds.jose.jwk.JWKSet
-import com.nimbusds.jose.jwk.KeyUse
-import com.nimbusds.jose.jwk.gen.ECKeyGenerator
 import com.nimbusds.jose.util.Base64URL
 import eu.europa.ec.eudi.openid4vp.*
 import eu.europa.ec.eudi.openid4vp.internal.AbsoluteDIDUrl
@@ -506,17 +502,3 @@ class RequestAuthenticatorOverDCApiTest {
         }
     }
 }
-//
-// Support
-//
-
-fun randomKey(): Pair<JWSAlgorithm, ECKey> =
-    JWSAlgorithm.ES256 to ECKeyGenerator(Curve.P_256).keyUse(KeyUse.SIGNATURE).generate()
-
-private inline fun <reified E : AuthorizationRequestError> assertFailsWithError(block: () -> Unit): E {
-    val exception = assertThrows<AuthorizationRequestException>(block)
-    return assertIs<E>(exception.error)
-}
-
-private fun UnvalidatedRequestObject.unsigned(): ReceivedRequest.Unsigned =
-    ReceivedRequest.Unsigned(this)
